@@ -7,18 +7,18 @@ import com.learn2code.Shop.db.service.api.request.UpdateProductRequest;
 import com.learn2code.Shop.domain.Customer;
 import com.learn2code.Shop.domain.Merchant;
 import com.learn2code.Shop.domain.Product;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import java.util.List;
 
-@RunWith(SpringRunner.class)
+@ExtendWith(SpringExtension.class)
 @SpringBootTest
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_CLASS)
 public class DBServiceTests {
@@ -34,7 +34,7 @@ public class DBServiceTests {
 
     private static Merchant merchant;
 
-    @Before
+    @BeforeEach
     public void createMerchant() {
         if (merchant == null) {
             merchant = new Merchant("name", "email", "address");
@@ -52,11 +52,11 @@ public class DBServiceTests {
         customer.setId(id);
 
         Customer fromDb = customerService.get(id);
-        Assert.assertEquals(customer, fromDb);
+        Assertions.assertEquals(customer, fromDb);
 
         List<Customer> customers = customerService.getCustomers();
-        Assert.assertEquals(1, customers.size());
-        Assert.assertEquals(customer, customers.get(0));
+        Assertions.assertEquals(1, customers.size());
+        Assertions.assertEquals(customer, customers.get(0));
     }
 
     @Test
@@ -64,11 +64,11 @@ public class DBServiceTests {
         // already created merchant in createMerchant function
 
         Merchant fromDB = merchantService.get(merchant.getId());
-        Assert.assertEquals(merchant, fromDB);
+        Assertions.assertEquals(merchant, fromDB);
 
         List<Merchant> merchants = merchantService.getMerchants();
-        Assert.assertEquals(1, merchants.size());
-        Assert.assertEquals(merchant, merchants.get(0));
+        Assertions.assertEquals(1, merchants.size());
+        Assertions.assertEquals(merchant, merchants.get(0));
     }
 
     @Test
@@ -79,21 +79,21 @@ public class DBServiceTests {
         product.setId(id);
 
         Product fromDB = productService.get(id);
-        Assert.assertEquals(product, fromDB);
+        Assertions.assertEquals(product, fromDB);
 
         List<Product> products = productService.getProducts();
-        Assert.assertEquals(1, products.size());
-        Assert.assertEquals(product, products.get(0));
+        Assertions.assertEquals(1, products.size());
+        Assertions.assertEquals(product, products.get(0));
 
         product.setAvailable(10);
         UpdateProductRequest productRequest = new UpdateProductRequest(product.getName(), product.getDescription(), product.getPrice(), product.getAvailable());
 
         productService.update(id, productRequest);
         Product fromDBAfterUpdate = productService.get(id);
-        Assert.assertEquals(product, fromDBAfterUpdate);
-        Assert.assertNotEquals(fromDB, fromDBAfterUpdate);
+        Assertions.assertEquals(product, fromDBAfterUpdate);
+        Assertions.assertNotEquals(fromDB, fromDBAfterUpdate);
 
         productService.delete(id);
-        Assert.assertEquals(0, productService.getProducts().size());
+        Assertions.assertEquals(0, productService.getProducts().size());
     }
 }
